--- a/benches/benches.rs
+++ b/benches/benches.rs
@@ -128,6 +128,40 @@ fn encode_decode_vec<T: TryFrom<u8> + Codec>(c: &mut Criterion) where T::Error: 
 	}, vec![1, 2, 5, 32, 1024, 2048, 16384]);
 }
 
+fn encode_decode_complex_type(c: &mut Criterion) {
+	#[derive(Encode, Decode, Clone)]
+	struct ComplexType {
+		_val: u32,
+		_other_val: u128,
+		_vec: Vec<u32>,
+	}
+
+	let complex_types = vec![
+		ComplexType { _val: 3, _other_val: 345634635, _vec: vec![1, 2, 3, 5, 6, 7] },
+		ComplexType { _val: 1000, _other_val: 0980345634635, _vec: vec![1, 2, 3, 5, 6, 7] },
+		ComplexType { _val: 43564, _other_val: 342342345634635, _vec: vec![1, 2, 3, 5, 6, 7] },
+	];
+	let complex_types2 = complex_types.clone();
+
+	c.bench_function_over_inputs("vec_encode_complex_type", move |b, &vec_size| {
+		let vec: Vec<ComplexType> = complex_types.clone().into_iter().cycle().take(vec_size).collect();
+
+		let vec = black_box(vec);
+		b.iter(|| vec.encode())
+	}, vec![1, 2, 5, 32, 1024, 2048, 16384]);
+
+	c.bench_function_over_inputs("vec_decode_complex_type", move |b, &vec_size| {
+		let vec: Vec<ComplexType> = complex_types2.clone().into_iter().cycle().take(vec_size).collect();
+
+		let vec = vec.encode();
+
+		let vec = black_box(vec);
+		b.iter(|| {
+			let _: Vec<ComplexType> = Decode::decode(&mut &vec[..]).unwrap();
+		})
+	}, vec![1, 2, 5, 32, 1024, 2048, 16384]);
+}
+
 fn bench_fn(c: &mut Criterion) {
 	c.bench_function("vec_write_as_output", vec_write_as_output);
 	c.bench_function("vec_extend", vec_extend);
@@ -176,6 +210,6 @@ criterion_group!{
 	config = Criterion::default().warm_up_time(Duration::from_millis(500)).without_plots();
 	targets = encode_decode_vec::<u8>, encode_decode_vec::<u16>, encode_decode_vec::<u32>, encode_decode_vec::<u64>,
 			encode_decode_vec::<i8>, encode_decode_vec::<i16>, encode_decode_vec::<i32>, encode_decode_vec::<i64>,
-			bench_fn, encode_decode_bitvec_u8
+			bench_fn, encode_decode_bitvec_u8, encode_decode_complex_type
 }
 criterion_main!(benches);