--- a/tests/mod.rs
+++ b/tests/mod.rs
@@ -489,7 +489,7 @@ fn codec_vec_u8() {
 		vec![0u8; 10],
 		vec![0u8; 100],
 		vec![0u8; 1000],
-	].into_iter() {
+	].iter() {
 		let e = v.encode();
 		assert_eq!(v, &Vec::<u8>::decode(&mut &e[..]).unwrap());
 	}