This pr brings some performance improvements, here are some benchmarks from before and after:

```
vec_decode_complex_type/1                                                                            
                        time:   [84.946 ns 85.588 ns 86.417 ns]
                        change: [-5.4537% -3.8573% -2.5559%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 7 outliers among 100 measurements (7.00%)
  2 (2.00%) high mild
  5 (5.00%) high severe
vec_decode_complex_type/2                                                                            
                        time:   [224.83 ns 225.17 ns 225.55 ns]
                        change: [+19.404% +19.821% +20.256%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 8 outliers among 100 measurements (8.00%)
  6 (6.00%) high mild
  2 (2.00%) high severe
vec_decode_complex_type/5                                                                            
                        time:   [483.57 ns 484.17 ns 484.89 ns]
                        change: [+19.389% +20.094% +20.636%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 8 outliers among 100 measurements (8.00%)
  3 (3.00%) high mild
  5 (5.00%) high severe
vec_decode_complex_type/32                                                                             
                        time:   [3.1287 us 3.1487 us 3.1872 us]
                        change: [+15.714% +16.734% +17.889%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 5 outliers among 100 measurements (5.00%)
  3 (3.00%) high mild
  2 (2.00%) high severe
vec_decode_complex_type/1024                                                                            
                        time:   [87.962 us 88.470 us 89.069 us]
                        change: [-91.906% -91.848% -91.796%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 10 outliers among 100 measurements (10.00%)
  6 (6.00%) high mild
  4 (4.00%) high severe
vec_decode_complex_type/2048                                                                            
                        time:   [175.26 us 175.80 us 176.46 us]
                        change: [-93.291% -93.222% -93.157%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 3 outliers among 100 measurements (3.00%)
  3 (3.00%) high severe
Benchmarking vec_decode_complex_type/16384: Warming up for 500.00 ms
Warning: Unable to complete 100 samples in 5.0s. You may wish to increase target time to 7.2s or reduce sample count to 50
vec_decode_complex_type/16384                                                                             
                        time:   [1.4089 ms 1.4114 ms 1.4149 ms]
                        change: [-95.519% -95.491% -95.451%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 8 outliers among 100 measurements (8.00%)
  5 (5.00%) high mild
  3 (3.00%) high severe
```

```
vec_encode_complex_type/1                                                                             
                        time:   [25.848 ns 26.059 ns 26.337 ns]
                        change: [-76.582% -76.372% -76.159%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 7 outliers among 100 measurements (7.00%)
  2 (2.00%) high mild
  5 (5.00%) high severe
vec_encode_complex_type/2                                                                             
                        time:   [40.648 ns 41.198 ns 41.896 ns]
                        change: [-82.797% -82.528% -82.175%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 6 outliers among 100 measurements (6.00%)
  2 (2.00%) high mild
  4 (4.00%) high severe
vec_encode_complex_type/5                                                                            
                        time:   [90.026 ns 90.806 ns 91.771 ns]
                        change: [-74.771% -74.517% -74.279%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 11 outliers among 100 measurements (11.00%)
  5 (5.00%) high mild
  6 (6.00%) high severe
vec_encode_complex_type/32                                                                            
                        time:   [477.59 ns 479.44 ns 481.54 ns]
                        change: [-68.468% -68.313% -68.129%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 7 outliers among 100 measurements (7.00%)
  5 (5.00%) high mild
  2 (2.00%) high severe
vec_encode_complex_type/1024                                                                             
                        time:   [13.408 us 13.443 us 13.480 us]
                        change: [-58.813% -58.468% -58.196%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 4 outliers among 100 measurements (4.00%)
  2 (2.00%) high mild
  2 (2.00%) high severe
vec_encode_complex_type/2048                                                                             
                        time:   [26.874 us 26.974 us 27.078 us]
                        change: [-58.554% -58.289% -58.049%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 2 outliers among 100 measurements (2.00%)
  2 (2.00%) high severe
vec_encode_complex_type/16384                                                                            
                        time:   [213.28 us 214.22 us 215.24 us]
                        change: [-59.009% -58.632% -58.145%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 5 outliers among 100 measurements (5.00%)
  4 (4.00%) high mild
  1 (1.00%) high severe
```

```
vec_decode_u32/1        time:   [42.708 ns 43.002 ns 43.511 ns]                              
                        change: [+54.739% +56.225% +57.650%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 4 outliers among 100 measurements (4.00%)
  2 (2.00%) high mild
  2 (2.00%) high severe
vec_decode_u32/2        time:   [64.711 ns 64.951 ns 65.214 ns]                             
                        change: [+112.99% +113.91% +114.82%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 4 outliers among 100 measurements (4.00%)
  3 (3.00%) high mild
  1 (1.00%) high severe
vec_decode_u32/5        time:   [64.485 ns 64.642 ns 64.806 ns]                             
                        change: [+52.461% +53.374% +54.169%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 2 outliers among 100 measurements (2.00%)
  1 (1.00%) high mild
  1 (1.00%) high severe
vec_decode_u32/32       time:   [80.770 ns 81.045 ns 81.348 ns]                              
                        change: [-50.849% -50.608% -50.360%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 4 outliers among 100 measurements (4.00%)
  2 (2.00%) high mild
  2 (2.00%) high severe
vec_decode_u32/1024     time:   [190.49 ns 191.91 ns 193.73 ns]                                
                        change: [-95.503% -95.475% -95.446%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 9 outliers among 100 measurements (9.00%)
  6 (6.00%) high mild
  3 (3.00%) high severe
vec_decode_u32/2048     time:   [356.93 ns 358.87 ns 361.29 ns]                                
                        change: [-95.914% -95.889% -95.862%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 8 outliers among 100 measurements (8.00%)
  4 (4.00%) high mild
  4 (4.00%) high severe
vec_decode_u32/16384    time:   [2.8900 us 2.8974 us 2.9052 us]                                  
                        change: [-95.849% -95.818% -95.791%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 3 outliers among 100 measurements (3.00%)
  1 (1.00%) high mild
  2 (2.00%) high severe
```

```
vec_decode_u8/1         time:   [32.216 ns 32.307 ns 32.407 ns]                             
                        change: [+4.3460% +4.7126% +5.0877%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 3 outliers among 100 measurements (3.00%)
  2 (2.00%) high mild
  1 (1.00%) high severe
vec_decode_u8/2         time:   [53.549 ns 53.641 ns 53.741 ns]                            
                        change: [+2.8426% +3.2120% +3.5870%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 2 outliers among 100 measurements (2.00%)
  1 (1.00%) high mild
  1 (1.00%) high severe
vec_decode_u8/5         time:   [53.720 ns 53.907 ns 54.125 ns]                            
                        change: [+5.1186% +5.8970% +6.8267%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 1 outliers among 100 measurements (1.00%)
  1 (1.00%) high mild
vec_decode_u8/32        time:   [54.240 ns 54.536 ns 55.009 ns]                             
                        change: [+2.8325% +3.3229% +3.9889%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 4 outliers among 100 measurements (4.00%)
  3 (3.00%) high mild
  1 (1.00%) high severe
vec_decode_u8/1024      time:   [105.29 ns 105.92 ns 106.77 ns]                               
                        change: [+11.231% +11.764% +12.445%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 5 outliers among 100 measurements (5.00%)
  1 (1.00%) high mild
  4 (4.00%) high severe
vec_decode_u8/2048      time:   [126.42 ns 127.47 ns 128.61 ns]                               
                        change: [-14.818% -14.225% -13.562%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 11 outliers among 100 measurements (11.00%)
  7 (7.00%) high mild
  4 (4.00%) high severe
vec_decode_u8/16384     time:   [479.35 ns 480.74 ns 482.31 ns]                                
                        change: [-15.482% -15.171% -14.841%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 4 outliers among 100 measurements (4.00%)
  3 (3.00%) high mild
  1 (1.00%) high severe
```

```
vec_encode_u8/1         time:   [15.933 ns 15.973 ns 16.015 ns]                             
                        change: [-1.5907% -1.0646% -0.5363%] (p = 0.00 < 0.05)
                        Change within noise threshold.
Found 7 outliers among 100 measurements (7.00%)
  7 (7.00%) high mild
vec_encode_u8/2         time:   [15.850 ns 15.887 ns 15.929 ns]                             
                        change: [-0.1073% +0.3561% +0.8222%] (p = 0.13 > 0.05)
                        No change in performance detected.
Found 4 outliers among 100 measurements (4.00%)
  2 (2.00%) high mild
  2 (2.00%) high severe
vec_encode_u8/5         time:   [16.296 ns 16.367 ns 16.469 ns]                             
                        change: [-3.7299% -2.5045% -1.4604%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 8 outliers among 100 measurements (8.00%)
  4 (4.00%) high mild
  4 (4.00%) high severe
vec_encode_u8/32        time:   [16.214 ns 16.247 ns 16.282 ns]                              
                        change: [-0.7405% -0.3667% -0.0155%] (p = 0.05 < 0.05)
                        Change within noise threshold.
Found 3 outliers among 100 measurements (3.00%)
  2 (2.00%) high mild
  1 (1.00%) high severe
vec_encode_u8/1024      time:   [23.774 ns 23.859 ns 23.956 ns]                                
                        change: [-5.9194% -5.3765% -4.8491%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 4 outliers among 100 measurements (4.00%)
  3 (3.00%) high mild
  1 (1.00%) high severe
vec_encode_u8/2048      time:   [67.023 ns 67.239 ns 67.496 ns]                               
                        change: [-15.965% -15.245% -14.566%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 8 outliers among 100 measurements (8.00%)
  4 (4.00%) high mild
  4 (4.00%) high severe
vec_encode_u8/16384     time:   [362.63 ns 363.81 ns 365.23 ns]                                
                        change: [-1.6168% -0.8091% -0.1145%] (p = 0.04 < 0.05)
                        Change within noise threshold.
Found 6 outliers among 100 measurements (6.00%)
  5 (5.00%) high mild
  1 (1.00%) high severe
```

```
vec_encode_u32/1        time:   [17.435 ns 17.607 ns 17.820 ns]                              
                        change: [-44.685% -43.810% -42.968%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 9 outliers among 100 measurements (9.00%)
  5 (5.00%) high mild
  4 (4.00%) high severe
vec_encode_u32/2        time:   [17.533 ns 17.674 ns 17.822 ns]                              
                        change: [-73.356% -73.125% -72.874%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 5 outliers among 100 measurements (5.00%)
  4 (4.00%) high mild
  1 (1.00%) high severe
vec_encode_u32/5        time:   [17.975 ns 18.083 ns 18.198 ns]                              
                        change: [-86.811% -86.599% -86.334%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 9 outliers among 100 measurements (9.00%)
  4 (4.00%) high mild
  5 (5.00%) high severe
vec_encode_u32/32       time:   [20.431 ns 20.517 ns 20.628 ns]                               
                        change: [-93.605% -93.573% -93.539%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 13 outliers among 100 measurements (13.00%)
  5 (5.00%) low mild
  2 (2.00%) high mild
  6 (6.00%) high severe
vec_encode_u32/1024     time:   [83.495 ns 84.623 ns 86.429 ns]                                
                        change: [-98.174% -98.140% -98.107%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 5 outliers among 100 measurements (5.00%)
  4 (4.00%) high mild
  1 (1.00%) high severe
vec_encode_u32/2048     time:   [177.84 ns 178.66 ns 179.59 ns]                                
                        change: [-97.914% -97.880% -97.852%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 3 outliers among 100 measurements (3.00%)
  2 (2.00%) high mild
  1 (1.00%) high severe
vec_encode_u32/16384    time:   [1.5940 us 1.6017 us 1.6098 us]                                  
                        change: [-97.412% -97.393% -97.373%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 12 outliers among 100 measurements (12.00%)
  8 (8.00%) high mild
  4 (4.00%) high severe
````