@thiolliere 2020-02-19
code is all correct to me, maybe we could add more test to test specialized implementation of some types but could be ok as is.